requirements.txt
Django==4.2.7
Pillow==10.1.0
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
python-decouple==3.8
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-filter==23.5
django-cors-headers==4.3.1
